+++ src/main/scala/com/raisondata/openai/api/audio/Transcription.scala
@@ -1,5 +1,6 @@
-package com.raisondata.openai.audio
+package com.raisondata.openai.api.audio
 
+import com.raisondata.openai.Language.Language
 import com.raisondata.openai.Model.Model
 import com.raisondata.openai.ResponseFormat.ResponseFormat
 import com.raisondata.openai._
@@ -10,16 +11,16 @@ import zio._
 
 import java.io.File
 
-object Translation extends SttpConfig with AudioMarshaller {
+object Transcription extends SttpConfig with AudioMarshaller {
+  override def usage: String = "transcriptions"
   override def domain: String = "audio"
 
-  override def usage: String = "translations"
-
-  def translate(
+  def transcribe(
       filePath: String,
       model: Model,
       responseFormat: ResponseFormat,
       temperature: Int,
+      language: Language,
       prompt: Option[String] = None
   )(openaiAPIKey: String): ZIO[Any, Throwable, TextResponse] =
     HttpClientZioBackend().flatMap { backend =>
@@ -40,22 +41,27 @@ object Translation extends SttpConfig with AudioMarshaller {
           "response_format",
           ResponseFormat.parse(responseFormat)
         ),
-        multipart("temperature", temperature)
+        multipart("temperature", temperature),
+        multipart(
+          "language",
+          Language.parse(language)
+        )
       )
 
       val request =
         requestWithMultipartForm(body, asJson[TextResponse])(openaiAPIKey)
       val response = getResponse(request)(backend)
 
-      response.map(_.body match {
-        case Left(error) =>
-          println(s"An error occurred while making a request $error")
-          throw new RuntimeException(error)
-        case Right(value) =>
-          println(s"Audio was transcribed successfully!")
-          println(value)
-          value
-      })
-
+      response
+        .map(_.body match {
+          case Left(error) =>
+            println(s"An error occurred while making a request $error")
+            throw new RuntimeException(error)
+          case Right(value) =>
+            println(s"Audio was transcribed successfully!")
+            println(value)
+            value
+        })
     }
+
 }
