+++ src/main/scala/com/raisondata/openai/api/images/EditImage.scala
