+++ src/main/scala/com/raisondata/openai/api/images/GenerateImage.scala
