+++ src/main/scala/com/raisondata/openai/api/images/ImageMarshaller.scala
