+++ src/main/scala/com/raisondata/openai/api/images/ImageVariation.scala
